% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{A table function to use with magrittr pipes}
\usage{
categorize(.data, ...)
}
\arguments{
\item{.data}{A data frame}

\item{...}{A formula (as in xtabs) or one or more objects which can be interpreted as factors (including character strings), or a list (or data frame) whose components can be so interpreted.}
}
\value{
A table (possibly as an xtabs class if a model formula was used)
}
\description{
Accepts a data frame as input and computes a contingency table for direct use in combination with the magrittr package.
}
\details{
categorize is a wrapper to xtabs or table such that a data frame can be given as the first argument.
}
\examples{

if (requireNamespace("magrittr", quietly = TRUE)) {
    library(magrittr)

    esoph \%>\% categorize(alcgp, agegp)
    esoph \%>\% categorize(~ alcgp + agegp)
}

}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{manip}
