% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipit.R
\name{clipit}
\alias{clipit}
\title{Copy an object as R-code to the clipboard}
\usage{
clipit(x)
}
\arguments{
\item{x}{object to copy}
}
\value{
Nothing but will put the R object into the clipboard as a side effect
}
\description{
Copies an R object to the clipboard so it can be pasted in elsewhere.
}
\details{
Returns nothing but will place the object in the clipboard
}
\examples{

clipit(mtcars$mpg)

}
\author{
Jonas LindeLøv posted on twitter \url{https://twitter.com/jonaslindeloev/status/1539182627554570240}. Copied shamelessly by Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{datagen}
