% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfunctions.R
\name{col.alpha}
\alias{col.alpha}
\title{Add and set alpha channel for RGB color}
\usage{
col.alpha(col, alpha = 1)
}
\arguments{
\item{col}{a vector of RGB color(s)}

\item{alpha}{numeric value between 0 and 1. Zero results fully transparent and 1 means full opacity}
}
\description{
Add and set alpha channel
}
\details{
This function adds and set an alpha channel to a RGB color
}
\examples{

newcol <- col.alpha("blue", .5)

}
\references{
Ekstrom, CT (2011) \emph{The R Primer}.
}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{iplot}
