% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfunctions.R
\name{col.tint}
\alias{col.tint}
\title{Tint an RGB color}
\usage{
col.tint(x, tint = 0.5)
}
\arguments{
\item{x}{an RGB color}

\item{tint}{numeric value between 0 and 1. Zero results in white and 0 means no change}
}
\description{
Tints an RBG color
}
\details{
This function tints an RGB color and returns the tinted RGB color (with alpha channel added)
}
\examples{

newcol <- col.tint("blue")

}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\references{
Ekstrom, CT (2011) \emph{The R Primer}.
}
\keyword{iplot}

