% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colCumSum}
\alias{colCumSum}
\title{Apply cumsum to each column of matrix}
\usage{
colCumSum(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A matrix the same size as m with the column-wise cumulative sums.
}
\description{
Fast computation of apply(x,2,cumsum)
}
\examples{
  // Generate 100000 predictors and 100 observations
  x <- matrix(rnorm(100*100000), nrow=100)
  colCumSum(y, x)

}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
