% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dCor}
\alias{dCor}
\title{Fast distance correlation matrix}
\usage{
dCor(x, y)
}
\arguments{
\item{x}{A matrix with dimensions n*k.}

\item{y}{A matrix with dimensions n*l.}
}
\value{
A number between 0 and 1 representing the distance covariance between x and y
}
\description{
Fast computation of the distance correation matrix between two matrices with the same number of rows. Note that this is not the same as the correlation matrix distance that can be computed with the cmd function.
}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
