% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dCov}
\alias{dCov}
\title{Fast distance covariance matrix}
\usage{
dCov(x, y)
}
\arguments{
\item{x}{A matrix with dimensions n*k.}

\item{y}{A matrix with dimensions n*l.}
}
\value{
A number representing the distance covariance between x and y
}
\description{
Fast computation of the distance covariance between two matrices with the same number of rows.
}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
