% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/desctable.R
\name{desctable}
\alias{desctable}
\alias{desctable.character}
\alias{desctable.data.frame}
\alias{desctable.default}
\alias{desctable.factor}
\alias{desctable.integer}
\alias{desctable.numeric}
\title{Create a descriptive table of summaries}
\usage{
desctable(object, group = NULL, proportions = FALSE, ...)
}
\arguments{
\item{object}{an object for which a descriptive summary table is desired}

\item{group}{A factor the same length as object (or the elements of object
if that is a data frame) that is used to split up the summary for each level
of group. If set to NULL (the default) then every observation is assumed to
be in one group}

\item{proportions}{Logical. Should proportions be listed for categorical
data (defaults to FALSE)}

\item{\dots}{any other arguments passed of functions}
}
\value{
A list with length the same as object with items \item{output
}{Output for use with the descriptive table} \item{p }{p-value for testing
equality across groups (set to NULL if groups=NULL)}
}
\description{
Creates a descriptive table of summary statistics
}
\examples{
desctable(ToothGrowth, ToothGrowth$supp)
}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\seealso{
\code{\link{summary}}
}
\keyword{print}

