% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Expand matrices or tables}
\usage{
expand(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A data frame with the matrix expanded
}
\description{
Compute power of test, or determine parameters to obtain target power.
}
\details{
The procedure uses uniroot to find the root of a discontinuous function so
some errors may pop up due to the given setup that causes the root-finding
procedure to fail. Also, since exact binomial tests are used we have
discontinuities in the function that we use to find the root of but despite
this the function is usually quite stable.
}
\examples{


}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{manip}

