% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_matrix}
\alias{expand_matrix}
\title{Expand matrix or table to data frame}
\usage{
expand_matrix(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
A data frame with the matrix expanded
}
\description{
Expands a contingency table to become a data frame where each observation becomes a combination of rows and columns.
}
\details{
The procedure uses uniroot to find the root of a discontinuous function so
some errors may pop up due to the given setup that causes the root-finding
procedure to fail. Also, since exact binomial tests are used we have
discontinuities in the function that we use to find the root of but despite
this the function is usually quite stable.
}
\examples{

expand_matrix(diag(3))
m <- matrix(c(2, 1, 3, 0, 0, 2), 3)
expand_matrix(m)
}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{manip}

