% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac2num.R
\name{fac2num}
\alias{fac2num}
\title{Convert factor to numeric vector}
\usage{
fac2num(x)
}
\arguments{
\item{x}{A factor}
}
\value{
Returns a numeric vector of the same length as x
}
\description{
Converts the factor labels to a vector of numeric vectors
}
\details{
Returns a vector of numeric values. Elements in the input factor that cannot be converted to numeric will produce NA.
}
\examples{

m1 <- matrix(rep(1, 16), 4)
m2 <- matrix(c(1, 0, .5, .5, 0, 1, .5, .5, .5, .5, 1, .5, .5, .5, .5, 1), 4)
m3 <- matrix(c(1, 1, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1), 4)
cmd(m1, m1)
cmd(m1, m2)
cmd(m2, m3)

}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{manip}

