% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filldown}
\alias{filldown}
\title{Fill down NA with the last observed observation}
\usage{
filldown(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector or list with the NA's replaced by the last observed value.
}
\description{
Fill down missing values with the latest non-missing value
}
\examples{

a <- c(1:5, "Howdy", NA, NA, 2:3, NA)
filldown(a)
filldown(c(NA, NA, NA, 3:5))

}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
