% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc.R
\name{hcc}
\alias{hcc}
\title{Compute Hajnal-Curie-Cohen model}
\usage{
hcc(limit = 12, S = 1, l = 0.86, A = 60000, Q = 6/7, CC = 4,
  agediff = setNames(rep(1, 41), -20:20), FUN.nat.age = hcc.nat.age,
  FUN.aid.age = hcc.aid.age, FUN.aid.fam.age = hcc.aid.family.age,
  NaturalChildren = c(0.2, 0.3, 0.3, 0.2), malefreq = 0.51,
  familyfreq = 0.14, criteria = c("offspring", "families"))
}
\arguments{
\item{limit}{the number of AID children for each donor}

\item{S}{the number of donors}

\item{NaturalChildren}{A vector of numbers representing the probabilities of that number of natural children (first element corresponds to zero children, second to 1 etc.).}

\item{malefreq}{the frequency of male births (defaults to 0.51). Set to NULL to force equal male/female distribution among donor offspring}

\item{diag}{logical. Should the diagonal be included?}
}
\value{
Returns a numeric vector containing the elements of the lower
triangular sub matrices.
}
\description{
XXXSplit a matrix into block diagonal sub matrices according to clusters and
XXXcombine the lower triangular parts into a vector
}
\examples{

m <- matrix(1:64, ncol=8)
cluster <- c(1, 1, 1, 1, 2, 2, 3, 3)
lower.tri.vector(m, cluster)

}
\author{
Claus Ekstrom \email{claus@ekstroem.dk}
}
\seealso{
\code{\link{lower.tri}}
}
\keyword{manip}

