% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ks_cumtest}
\alias{ks_cumtest}
\title{Kolmogorov-Smirnov goodness of fit test for cumulative discrete data}
\usage{
ks_cumtest(x, B = 10000L, prob = NULL)
}
\arguments{
\item{x}{A vector representing the contingency table.}

\item{B}{The number of simulations used to compute the p-value.}

\item{prob}{A positive vector of the same length as x representing the distribution under the null hypothesis. It will be scaled to sum to 1. If NULL (the default) then a uniform distribution is assumed.}
}
\value{
A list of class "htest" giving the simulation results.
}
\description{
Kolmogorov-Smirnov goodness of fit test for cumulative discrete data.
}
\details{
The name of the function might change in the future so keep that in mind!

Simulation is done by random sampling from the null hypothesis.
}
\examples{

x <- 1:6
ks_cumtest(x)

}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
