% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESS-package.R
\docType{data}
\name{kwdata}
\alias{kwdata}
\title{Non-parametric Kruskal Wallis data example}
\format{
An artificial data frame with 18 observations in each of three groups.
\describe{ \item{x}{measurements for group 1}
\item{y}{measurements for group 2}
\item{z}{measurements for group 3}}
}
\source{
Data example found on the internet
}
\description{
Artificial dataset to show that the p-value obtained for the Kruskal Wallis is only valid _after_ the distributional form
has been checked to be the same for all groups.
}
\examples{

data(kwdata)
newdata <- stack(kwdata)
kruskal.test(values ~ ind, newdata)

}
\keyword{datasets}
