% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESS-package.R
\docType{data}
\name{lifeexpect}
\alias{lifeexpect}
\title{Estimated life expectancy for Danish newborns}
\format{
A data frame with 70 observations on the following 3 variables.
  \describe{
    \item{\code{year}}{a character vector} giving the calendar interval on which the estimation was based.
    \item{\code{male}}{a numeric vector} Life expectancy for males (in years).
    \item{\code{female}}{a numeric vector} Life expectancy for females (in years)
    \item{\code{myear}}{a numeric vector} The midpoint of the year interval
  }
}
\source{
Data collected from Danmarsk Statistik.
See \url{https://www.dst.dk/en} for more information.
}
\description{
The estimated life expectancy for newborn Danes split according to gender.
}
\examples{

data(lifeexpect)
plot(lifeexpect$myear, lifeexpect$male)


}
\keyword{datasets}
