% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmm_Maximize_cpp}
\alias{lmm_Maximize_cpp}
\title{Maximize linear mixed model with user-specified covariance matrices}
\usage{
lmm_Maximize_cpp(y, x, vc, maxiter = 25L, REML = TRUE, tolerance = 1e-06,
  reparam = FALSE, scale = TRUE)
}
\arguments{
\item{y}{A vector of outcomes.}

\item{x}{A design matrix of regressor variables. Must have the same number of rows as the length of y.}

\item{vc}{A list of fixed variance-covariance matrices}

\item{maxiter}{Maximum number of iterations}

\item{REML}{A logical that determines if restricted maximum likelihood (REML - the default) or maximum likelihood (ML)}

\item{tolerance}{The Maximum number of iterations}
}
\value{
A data frame with two variables: coefficients and stderr that gives the slope estimate and corresponding standard error for each column in x.
}
\description{
Fast computation of simple regression slopes for each predictor represented by a column in a matrix
}
\author{
Claus Ekstrom <claus@rprimer.dk>
}

