% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Load and extract object from RData file}
\usage{
loadRData(fileName)
}
\arguments{
\item{filename}{The path to the RData file}
}
\value{
An R object
}
\description{
Loads and extracts an object from an RData file
}
\details{
Returns an R object
}
\examples{

\dontrun{
  d <- loadRData("~/blah/ricardo.RData")
}

}
\seealso{
\code{\link{load}}
}
\author{
ricardo (from GitHub)
}
\keyword{manip}
