% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESS-package.R
\docType{data}
\name{matched}
\alias{matched}
\title{Flu hospitalization}
\format{
A data frame with 450 observations on the following 4 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{iscase}}{a factor with levels \code{Control} \code{Case}}
    \item{\code{vaccine}}{a factor with levels \code{Not} \code{Vaccinated}}
    \item{\code{lung}}{a factor with levels \code{None} \code{Disease}}
  }
}
\source{
Modified from: Stokes, Davis, Koch (2000). "Categorical Data Analysis Using the SAS System," Chapter 10.
}
\description{
Researchers in a Midwestern county tracked flu cases requiring hospitalization in those residents aged 65
and older during a two-month period one winter. They matched each case with 2 controls by sex and age (150 cases,
300 controls). They used medical records to determine whether cases and controls had received a flu vaccine shot
and whether they had underlying lung disease. They wanted to know whether flu vaccination prevents hospitalization
for flu (severe cases of flu). Underlying lung disease is a potential confounder.
}
\examples{

data(matched)

}
\keyword{datasets}
