% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maximum_subarray}
\alias{maximum_subarray}
\title{Fast computation of maximum sum subarray}
\usage{
maximum_subarray(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A list with three elements: sum (the maximum subarray sum), start (the starting index of the subarray) and end (the ending index of the subarray)
}
\description{
Fast computation of the maximum subarray sum of a vector using Kadane's algorithm. The implementation handles purely negative numbers.
}
\examples{

maximum_subarray(1:4)

maximum_subarray(c(-2, 1, -3, 4, -1, 2, 1, -5, 4))
 
maximum_subarray(rnorm(100000))

}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
