\name{MESS-package}
\alias{MESS-package}
\alias{MESS}
\docType{package}
\title{
Collection of miscellaneous useful and semi-useful functions
}
\description{
Collection of miscellaneous useful and semi-useful functions and add-on
functions that enhances a number of existing packages and provides
In particular in relation to statistical genetics
}
\details{
\tabular{ll}{
Package: \tab MESS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-03-29\cr
License: \tab GPL-2\cr
}
% ~~ An overview of how to use the package, including the most important ~~
% ~~ functions ~~
}
\author{
Claus Ekstrom \email{claus@rprimer.dk}\cr
Maintainer: Claus Ekstrom \email{claus@rprimer.dk}
}
\references{
Ekstrom, C. (2011). The R Primer. Chapman & Hall.
}
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
