% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESS-package.R
\docType{package}
\name{MESS}
\alias{MESS}
\alias{MESS-package}
\title{Collection of miscellaneous useful and semi-useful functions}
\description{
Collection of miscellaneous useful and semi-useful functions and add-on
functions that enhances a number of existing packages and provides In
particular in relation to statistical genetics
}
\details{
\tabular{ll}{ Package: \tab MESS\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2012-03-29\cr License: \tab GPL-2\cr } % ~~ An overview of
how to use the package, including the most important ~~ % ~~ functions ~~
}
\references{
Ekstrøm, C. (2011). The R Primer. Chapman & Hall.
}
\author{
Claus Thorn Ekstrøm \email{claus@rprimer.dk}\cr Maintainer: Claus Thorn Ekstrøm
\email{claus@rprimer.dk}
}
\keyword{package}
