% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mfastLm_cpp}
\alias{mfastLm_cpp}
\title{Fast marginal simple regresion analyses}
\usage{
mfastLm_cpp(y, x, addintercept)
}
\arguments{
\item{y}{A vector of outcomes.}

\item{x}{A matrix of regressor variables. Must have the same number of rows as the length of y.}

\item{addintercept}{A logical that determines if the intercept should be included in all analyses (TRUE) or not (FALSE)}
}
\value{
A data frame with two variables: coefficients and stderr that gives the slope estimate and corresponding standard error for each column in x.
}
\description{
Fast computation of simple regression slopes for each predictor represented by a column in a matrix
}
\author{
Claus Ekstrøm <claus@rprimer.dk>
}

