% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mrbind}
\alias{mrbind}
\title{Fast replication of a matrix}
\usage{
mrbind(m1, m2)
}
\arguments{
\item{m1}{A matrix with dimensions r*c.}

\item{m2}{An integer giving the number of times the matrix is replicated row-wise}
}
\value{
A matrix with dimensions (r*nrow) x (c*ncol)
}
\description{
Fast generation of a matrix by replicating a matrix row- and column-wise in a block-like fashion
}
\author{
Claus Ekstrom <claus@rprimer.dk>
}

