% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESS-package.R
\docType{data}
\name{nh4}
\alias{nh4}
\title{Ammonia nitrogen found in river}
\format{
A data frame with 120 observations on the following 3 variables.
\describe{\item{nh4}{The ammonia nitrogen levels (mg/l). A value of zero corresponds to a censoring, but it really is censored at <0.01}
\item{cens}{A logical vector indicating if the value was censored}
\item{year}{The year}}
}
\source{
Found on the internet and partly simulated
}
\description{
Monthly levels of ammonia nitrogen in a river over two years
}
\examples{

data(nh4)

}
\keyword{datasets}
