% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered.clusters.R
\name{ordered.clusters}
\alias{ordered.clusters}
\title{Check if unique elements of a vector appear in contiguous clusters}
\usage{
ordered.clusters(id)
}
\arguments{
\item{id}{a vector}
}
\value{
The function returns TRUE if the elements appear in contiguous
clusters and FALSE otherwise
}
\description{
\code{ordered.clusters} determines if identical elements of a vector appear
in contiguous clusters, and returns \code{TRUE} if the do and \code{FALSE}
otherwise.
}
\examples{

x <- c(1, 1, 1, 2, 2, 3, 4, 1, 5, 5, 5)
ordered.clusters(x)
ordered.clusters(sort(x))
ordered.clusters(x[order(x)])

}
\seealso{
\code{\link{duplicated}}
}
\author{
Claus Ekstrom \email{claus@ekstroem.dk} with suggestions from Peter
Dalgaard.
}
\keyword{utilities}
