% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{page}
\alias{page}
\title{Fast replication of a matrix}
\usage{
page(from, to)
}
\arguments{
\item{x}{A matrix with dimensions r*c.}

\item{nrow}{An integer giving the number of times the matrix is replicated row-wise}

\item{ncol}{An integer giving the number of times the matrix is replicated column-wise}
}
\value{
A matrix with dimensions (r*nrow) x (c*ncol)
}
\description{
Fast generation of a matrix by replicating a matrix row- and column-wise in a block-like fashion
}
\author{
Claus Ekstrom <claus@rprimer.dk>
}

