% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qdiag}
\alias{qdiag}
\title{Fast extraction of matrix diagonal}
\usage{
qdiag(x)
}
\arguments{
\item{x}{The matrix to extract the diagonal from}
}
\value{
A vector with the diagonal elements
}
\description{
Fast extraction of matrix diagonal
}
\details{
Note this function can only be used for extraction
}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
