% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quadform}
\alias{quadform}
\title{Fast quadratic form computation}
\usage{
quadform(x, M, invertM = FALSE, transposex = FALSE)
}
\arguments{
\item{x}{A matrix with dimensions n*k.}

\item{M}{A matrix with dimenions n*n. If it is to be inverted then the matrix should be symmetric and positive difinite (no check is done for this)}

\item{invertM}{A logical. If set to TRUE then M will be inverted before computations (defaults to FALSE)}

\item{transposex}{A logical. Should the matrix be transposed before computations (defaults to FALSE).}
}
\value{
A matrix with dimensions k * k giving the quadratic form
}
\description{
Fast computation of a quadratic form  \eqn{t(x) * M * x}.
}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
