% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnormperfect.R
\name{rnorm_perfect}
\alias{rnorm_perfect}
\title{Simulate values from a perfect normal distribution}
\usage{
rnorm_perfect(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{mean}{number of mean.}

\item{sd}{number of standard deviation.}
}
\value{
Returns a vector of values from a perfect normal distribution
}
\description{
Random generation for a perfect normal distribution with mean equal to mean and standard deviation equal to sd.
}
\details{
The function will return the same set of quantiles for fixed n. In that sense there is not much randomness going on, and the function is mostly useful for illustrative purposes.
}
\examples{

rnorm_perfect(30, mean=10, sd=2)

}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{hplot}
