% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screener.R
\name{screen_variables}
\alias{screen_variables}
\title{Screen variable before penalized regression}
\usage{
screen_variables(x, y, lambda = 0, method = c("xxx"))
}
\arguments{
\item{x}{A table or matrix}

\item{y}{A vector out outcomes}

\item{lambda}{a vector of values penalized ....}

\item{method}{a string giving the method used for screening. Currently not used}
}
\value{
A list with two elements: lambda which
}
\description{
Expands a contingency table to a data frame where each observation in the table becomes a single observation in the data frame with corresponding information for each for each combination of the table dimensions.
}
\details{
Note that no standardization is done (not necessary?)
}
\examples{

x <- matrix(rnorm(50*100), nrow=50)
y <- rnorm(50)
screen_variables(x, y, lambda=c(0, 1, 2))

}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{manip}
