% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{select_variables_using_cor}
\alias{select_variables_using_cor}
\title{Fast selection of variables below correlation threshold}
\usage{
select_variables_using_cor(mat, threshold)
}
\arguments{
\item{threshold}{The threshold. Should be a number between 0 and 1.}

\item{x}{A matrix}
}
\value{
A logical vector where the elements that are TRUE correspond to the variables that are not correlated above the threshold with any of the subsequent variables.
}
\description{
I'll need to fill this out.
}
\examples{

x <- sample(10, 20, replace = TRUE)
bin(x, 15)

}
\author{
Claus Ekstrøm <claus@rprimer.dk>
}
