% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinv.R
\name{sinv}
\alias{sinv}
\title{Invert a symmetric positive-definite matrix}
\usage{
sinv(obj)
}
\arguments{
\item{obj}{The symmetric positive-definite matrix}
}
\value{
A matrix of the same size as the input object
}
\description{
Inverts a symmetric positive-definite matrix without requiring the Matrix package.
}
\details{
This function does no error checking and it is up to the user to ensure that the input is indeed symmetric, positive-definite, and a matrix.
}
\examples{
m <- matrix(c(1, 0, .5, .5, 0, 1, .5, .5, .5, .5, 1, .5, .5, .5, .5, 1), 4)
sinv(m)

}
\author{
Claus Ekstrom, \email{claus@rprimer.dk}.
}
\keyword{file}
