% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tracemp}
\alias{tracemp}
\title{Fast computation of trace of matrix product}
\usage{
tracemp(A, B)
}
\arguments{
\item{A}{A matrix with dimensions n*k.}

\item{B}{A matrix with dimenions n*k.}
}
\value{
The trace of the matrix product
}
\description{
Fast computation of the trace of the matrix product trace(t(A) %*% B)
}
\examples{

A <- matrix(1:12, ncol=3)
tracemp(A, A)

}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
