% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usd.R
\name{usd}
\alias{usd}
\title{Unbiased standard deviation}
\usage{
usd(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an R object but not a factor coercible to numeric by as.double(x)}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
A scalar
}
\description{
This function computes the unbiased standard deviation of the values in x. If \code{na.rm} is TRUE then missing values are removed before computation proceeds.
}
\details{
Like var this uses denominator n - 1.
The standard deviation of a length-one or zero-length vector is NA.
}
\examples{
sd(1:5)
usd(1:5)

}
