% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writexml.R
\name{write.xml}
\alias{write.xml}
\title{Write a data frame in XML format}
\usage{
write.xml(data, file = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{the data frame object to save}

\item{file}{the file name to be written to.}

\item{na.rm}{logical. Should missing values be discarded from the written XML file (defaults to FALSE)}
}
\value{
None
}
\description{
Writes the data frame to a file in the XML format.
}
\details{
This function requires the \pkg{XML} package to be installed to function properly.
}
\examples{
data(trees)
write.xml(trees, file="mydata.xml")

}
\author{
Claus Ekstrom, \email{claus@rprimer.dk} based on previous work by Duncan Temple Lang.
}
\keyword{file}

