% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfunctions.R
\name{col.shade}
\alias{col.shade}
\title{Shade an RGB color}
\usage{
col.shade(col, shade = 0.5)
}
\arguments{
\item{col}{a vector of RGB color(s)}

\item{shade}{numeric value between 0 and 1. Zero means no change and 1 results in black}
}
\description{
Shades an RBG color
}
\details{
This function shades an RGB color and returns the shaded RGB color (with alpha channel added)
}
\examples{

newcol <- col.shade("blue")

}
\references{
Ekstrom, CT (2011) \emph{The R Primer}.
}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{iplot}
