% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colCumSum}
\alias{colCumSum}
\title{Apply cumsum to each column of matrix}
\usage{
colCumSum(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A matrix the same size as m with the column-wise cumulative sums.
}
\description{
Fast computation of apply(m, 2, cumsum)
}
\examples{
  # Generate a 100 by 10000 matrix
  x <- matrix(rnorm(100*10000), nrow=100)
  result <- colCumSum(x)

}
\author{
Claus Ekstrom <claus@rprimer.dk>
}
