% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac2num.R
\name{fac2num}
\alias{fac2num}
\title{Convert factor to numeric vector}
\usage{
fac2num(x)
}
\arguments{
\item{x}{A factor}
}
\value{
Returns a numeric vector of the same length as x
}
\description{
Converts the factor labels to numeric values and returns the factor as a numeric vector
}
\details{
Returns a vector of numeric values. Elements in the input factor that cannot be converted to numeric will produce NA.
}
\examples{

f <- factor(c(1,2,1,3,2,1,2,3,1))
fac2num(f)

}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{manip}
