% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc.R
\name{hcc}
\alias{hcc}
\title{Compute Hajnal-Curie-Cohen model}
\usage{
hcc(limit = 12, Q = 6/7, malefreq = 0.51, familyfreq = 0,
  criteria = c("offspring", "families"))
}
\arguments{
\item{limit}{a square matrix}

\item{malefreq}{the frequency of male births (defaults to 0.51). Set to NULL to force equal male/female distribution among donor offspring}

\item{diag}{logical. Should the diagonal be included?}
}
\value{
Returns a numeric vector containing the elements of the lower
triangular sub matrices.
}
\description{
XXXSplit a matrix into block diagonal sub matrices according to clusters and
XXXcombine the lower triangular parts into a vector
}
\examples{

m <- matrix(1:64, ncol=8)
cluster <- c(1, 1, 1, 1, 2, 2, 3, 3)
lower.tri.vector(m, cluster)

}
\author{
Claus Ekstrom \email{claus@ekstroem.dk}
}
\seealso{
\code{\link{lower.tri}}
}
\keyword{manip}

