% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcc.R
\name{hcc}
\alias{hcc}
\title{Compute Hajnal-Curie-Cohen model}
\usage{
hcc(limit = 12, S = 1, l = 0.86, A = 60000, Q = 6/7, CC = 4,
  agediff = setNames(c(1.6e-06, 1.6e-06, 1.6e-06, 7.8e-06, 1.57e-05, 1.72e-05,
  3.92e-05, 7.83e-05, 0.0001332, 0.0001864, 0.0002773, 0.0004982, 0.0007113,
  0.0011828, 0.0018972, 0.0027981, 0.0046013, 0.0070798, 0.0114273, 0.0178694,
  0.0279525, 0.0451279, 0.0751312, 0.1178276, 0.1257393, 0.1151502, 0.0992046,
  0.0802933, 0.0629127, 0.0480795, 0.0367776, 0.0277645, 0.021161, 0.0161132,
  0.0119725, 0.0090146, 0.0069952, 0.0052374, 0.0041564, 0.0031286, 0.002386,
  0.0019317, 0.0014868, 0.0011703, 0.0009807, 0.0006737,      0.000622,
  0.0005013, 0.0003885, 0.0002632, 0.0002225, 0.0001661, 0.0001379, 0.0001018,
  9.24e-05, 7.05e-05, 6.89e-05, 3.45e-05, 2.35e-05, 3.92e-05, 1.41e-05,
  1.41e-05, 1.1e-05, 4.7e-06, 3.1e-06, 9.4e-06, 3.1e-06, 4.7e-06, 3.1e-06,
  3.1e-06, 3.1e-06), -23:47), FUN.nat.age = hcc.nat.age,
  FUN.aid.age = hcc.aid.age, FUN.aid.fam.age = hcc.aid.family.age,
  NaturalChildren = c(0.2, 0.3, 0.3, 0.2), malefreq = 0.51,
  familyfreq = 0.14, criteria = c("offspring", "families"))
}
\arguments{
\item{limit}{the number of AID children for each donor}

\item{S}{the number of donors}

\item{NaturalChildren}{A vector of numbers representing the probabilities of that number of natural children (first element corresponds to zero children, second to 1 etc.).}

\item{malefreq}{the frequency of male births (defaults to 0.51). Set to NULL to force equal male/female distribution among donor offspring}

\item{diag}{logical. Should the diagonal be included?}
}
\value{
Returns a numeric vector containing the elements of the lower
triangular sub matrices.
}
\description{
XXXSplit a matrix into block diagonal sub matrices according to clusters and
XXXcombine the lower triangular parts into a vector
}
\examples{

m <- matrix(1:64, ncol=8)
cluster <- c(1, 1, 1, 1, 2, 2, 3, 3)
lower.tri.vector(m, cluster)

}
\author{
Claus Ekstrom \email{claus@ekstroem.dk}
}
\seealso{
\code{\link{lower.tri}}
}
\keyword{manip}

