% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MESS-package.R
\docType{package}
\name{MESS-package}
\alias{MESS}
\alias{MESS-package}
\title{MESS: Miscellaneous Esoteric Statistical Scripts}
\description{
A mixed collection of useful and semi-useful diverse statistical functions, some of which may even be referenced in The R Primer book. See Ekstrøm, C. T. (2016). The R Primer. 2nd edition. Chapman & Hall.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ekstroem/MESS}
  \item Report bugs at \url{https://github.com/ekstroem/MESS/issues}
}

}
\author{
\strong{Maintainer}: Claus Thorn Ekstrøm \email{claus@rprimer.dk}

Other contributors:
\itemize{
  \item Niels Aske Lundtorp Olsen \email{nalo@dtu.dk} [contributor]
}

}
\keyword{internal}
