% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{panel.hist}
\alias{panel.hist}
\title{Panel plot of histogram and density curve}
\usage{
panel.hist(x, col.bar = "gray", ...)
}
\arguments{
\item{x}{a numeric vector of x values}

\item{col.bar}{the color of the bars}

\item{...}{options passed to hist}
}
\description{
Prints the histogram and corresponding density curve
}
\details{
This function prints a combined histogram and density curve for use with
the pairs function
}
\examples{

pairs(~ Ozone + Temp + Wind + Solar.R, data=airquality,
      lower.panel=panel.smooth, diag.panel=panel.hist,
      upper.panel=panel.r2)

}
\references{
Ekstrom, CT (2011) \emph{The R Primer}.
}
\author{
Claus Ekstrom \email{claus@rprimer.dk}
}
\keyword{iplot}
