% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writexml.R
\name{write.xml}
\alias{write.xml}
\title{Write a data frame in XML format}
\usage{
write.xml(data, file = NULL, collapse = TRUE)
}
\arguments{
\item{data}{the data frame object to save}

\item{file}{the file name to be written to.}

\item{collapse}{logical. Should the output file be collapsed to make it fill less? (Defaults to TRUE)}
}
\value{
None
}
\description{
Writes the data frame to a file in the XML format.
}
\details{
This function does not require the \pkg{XML} package to be installed to function properly.
}
\examples{

\dontrun{
data(trees)
write.xml(trees, file="mydata.xml")
}

}
\author{
Claus Ekstrom, \email{claus@rprimer.dk} based on previous work by Duncan Temple Lang.
}
\keyword{file}
